requirements.txt
flask
numpy 
opencv-python
opencv-python-headless
tensorflow
keras
scipy
Pillow
